[Refactor] 임시저장 api 한 파일로 정리
