--- src/apis/portfolio/fetchTempItem.controller.ts
@@ -1,13 +1,0 @@
-import supabase from '@/lib/supabaseClient'
-
-export const fetchTempPortfolioItem = async (id: string) => {
-  const { data, error } = await supabase
-    .from('TempPortfolio')
-    .select('*')
-    .eq('id', id)
-    .maybeSingle()
-
-  if (error) throw error
-
-  return data ?? null
-}
